==> 3)Differential_Expression_Analysis.Rmd <==
## Differential Expression Analysis with edgeR

### Setup and Library Loading

```{r}
library(edgeR)
library(limma)
library(Glimma)
library(gplots)
library(org.Mm.eg.db)
load("Robjects/preprocessing.Rdata")
```

### Recap of Pre-processing

The following steps summarize the data pre-processing and transformation conducted earlier:

```{r eval=FALSE}
seqdata <- read.delim("data/GSE60450_Lactation-GenewiseCounts.txt", stringsAsFactors = FALSE)
countdata <- seqdata[,-(1:2)]
rownames(countdata) <- seqdata[,1]
colnames(countdata) <- substr(colnames(countdata), 1, 7)
myCPM <- cpm(countdata)
thresh <- myCPM > 0.5
keep <- rowSums(thresh) >= 2
counts.keep <- countdata[keep,]
dgeObj <- DGEList(counts.keep)
dgeObj <- calcNormFactors(dgeObj)
sampleinfo <- read.delim("data/SampleInfo_Corrected.txt")
group <- paste(sampleinfo$CellType, sampleinfo$Status, sep = ".")
```

### Creating the Design Matrix

We create a design matrix to model the effects of cell type and status. The first model includes only main effects:

```{r}
group <- as.character(group)
type <- sapply(strsplit(group, ".", fixed = TRUE), function(x) x[1])
status <- sapply(strsplit(group, ".", fixed = TRUE), function(x) x[2])
design <- model.matrix(~ type + status)
design
```

### Data Exploration

An MDS plot provides an overview of sample relationships based on biological coefficient of variation (BCV):

```{r}
plotMDS(dgeObj, labels = group, cex = 0.75, xlim = c(-4, 5))
```

### Dispersion Estimation

Estimating dispersion is crucial for modeling variability:

```{r}
dgeObj <- estimateCommonDisp(dgeObj)
dgeObj <- estimateGLMTrendedDisp(dgeObj)
dgeObj <- estimateTagwiseDisp(dgeObj)
plotBCV(dgeObj)
```

### Testing for Differential Expression

We fit gene-wise GLMs and test for differential expression:

```{r}
fit <- glmFit(dgeObj, design)
lrt.BvsL <- glmLRT(fit, coef = 2)
topTags(lrt.BvsL)
```


### Contrasts

We construct contrasts to test specific hypotheses, such as differences between pregnant and virgin samples:

```{r}
PvsV <- makeContrasts(statuspregnant - statusvirgin, levels = design)
lrt.pVsV <- glmLRT(fit, contrast = PvsV)
topTags(lrt.pVsV)
```


### Saving Results

Save results for future use:

```{r}
save(lrt.BvsL, dgeObj, group, file = "Robjects/DE.Rdata")
